reorganize, move things around, delete old files
