+++ client/src/App.js
@@ -7,16 +7,13 @@ import categories from './categories.json';
 import SubNav from './components/subNav';
 import FlexContainer from './components/flexContainer';
 // import API from './utils/API';
+
 //Routes
-// import Home from './pages/Home';
 import Editor from './pages/Editor'
-import Content from './pages/Content';
 import NoMatch from './pages/NoMatch';
 // import * as routes from './constants/routes';
 import AccountPage from './pages/Account';
-// import MainCategoryPage from './pages/MainCategoryPage';
 import Profile from './pages/Profile';
-
 import authTest from './pages/AUTH-TEST';
 
 // Auth Helper
@@ -24,6 +21,7 @@ import withAuthentication from './components/AuthUserSession/withAuthentication'
 import ModalConductor from './components/Modals/ModalConductor'
 import MainPageContainer from './pages/Main/MainPageContainer';
 import PostDetailPage from './pages/Content/PostDetailPage';
+
 class App extends Component {
   constructor(props) {
     super(props);
@@ -60,13 +58,9 @@ class App extends Component {
             closeModal={this.closeModal}
           />
           <Switch>
-            {/* <Route exact path='/' component={Home} /> */}
             <Route exact path='/' component={MainPageContainer} />
-            {/* <Route exact path='/home' component={Home} /> */}
-            {/* <Route exact path='/categories/:categoryName' component={MainCategoryPage} /> */}
             <Route exact path='/categories/:categoryName' component={MainPageContainer} />
             <Route exact path='/categories/:categoryName/posts/new' component={Editor} />
-            {/* <Route exact path='/posts/:id' component={Content} /> */}
             <Route exact path='/posts/:postId' component={PostDetailPage} />
             <Route exact path='/posts/:postId/edit' component={Editor} />
             <Route exact path='/account' component={AccountPage} />
