--- client/src/components/PostComponents/PostDetail/index.js
@@ -1,1 +1,0 @@
-export * from "./PostDetail";
