+++ client/src/pages/Content/index.js
@@ -1,1 +1,1 @@
-export { default } from "./Content.js";
+export { default } from "./PostDetailPage";
