--- client/src/pages/MainCategoryPage/index.js
@@ -1,1 +1,0 @@
-export { default } from "./MainCategoryPage.js";
